+++ src/components/modules/SignIn/Main/index.tsx
@@ -1,10 +1,9 @@
 'use client';
 
-import { Flex } from 'antd';
 import Title from 'antd/es/typography/Title';
 
-import Typography from '@/components/core/common/Typography';
 import FormSignin from '@/components/modules/SignIn/FormSignin';
+import SignInSocial from '@/components/modules/SignInSocial';
 
 import * as S from './styles';
 
@@ -13,17 +12,7 @@ function SignIn() {
         <S.HomeWrapper>
             <Title level={2}>Đăng nhập</Title>
             <FormSignin/>
-            <Typography>
-                Đăng nhập bằng
-            </Typography>
-            <Flex gap={40}>
-                <S.ButtonLogo>
-                    <S.ImgLogo preview={false} alt="" src={'/images/signin/logo-facebook.png'} />
-                </S.ButtonLogo>
-                <S.ButtonLogo>
-                    <S.ImgLogo preview={false} alt="" src={'/images/signin/logo-gg.png'} />
-                </S.ButtonLogo>
-            </Flex>
+            <SignInSocial/>
         </S.HomeWrapper>
     )
 }
