+++ src/components/modules/SignInSocial/styles.ts
@@ -1,0 +1,23 @@
+import { Image } from "antd";
+import styled from "styled-components";
+
+export const ButtonLogo = styled.button`
+    position: relative;
+    
+    padding: 10px;
+
+    width: 70px;
+    height: 70px;
+
+    background-color: white;
+    border-radius: 10px;
+    
+    display: flex;
+    justify-content: center;
+    align-items: center;
+`
+
+export const ImgLogo = styled(Image)`
+  width: 100;
+  height: 100;
+`;
