+++ src/components/modules/SignUp/Main/index.tsx
@@ -1,10 +1,9 @@
 'use client';
 
-import { Flex } from 'antd';
 import Title from 'antd/es/typography/Title';
 
-import Typography from '@/components/core/common/Typography';
 import FormSignUp from '@/components/modules/SignUp/FormSignUp';
+import SignInSocial from '@/components/modules/SignInSocial';
 
 import * as S from './styles';
 function SignUp() {
@@ -12,17 +11,7 @@ function SignUp() {
         <S.HomeWrapper>
             <Title level={2}>Đăng ký</Title>
             <FormSignUp/>
-            <Typography>
-                Đăng nhập bằng
-            </Typography>
-            <Flex gap={40}>
-                <S.ButtonLogo>
-                    <S.ImgLogo preview={false} alt="" src={'/images/signin/logo-facebook.png'} />
-                </S.ButtonLogo>
-                <S.ButtonLogo>
-                    <S.ImgLogo preview={false} alt="" src={'/images/signin/logo-gg.png'} />
-                </S.ButtonLogo>
-            </Flex>
+            <SignInSocial/>
         </S.HomeWrapper>
     )
 }
