+++ src/components/modules/SignUp/Main/styles.ts
@@ -1,4 +1,4 @@
-import { Flex, Image } from 'antd';
+import { Flex } from 'antd';
 import styled from 'styled-components';
 
 export const HomeWrapper = styled(Flex)`
@@ -11,24 +11,4 @@ export const HomeWrapper = styled(Flex)`
 `;
 
 
-export const ButtonLogo = styled.button`
-    position: relative;
-
-    padding: 10px;
-
-    width: 70px;
-    height: 70px;
-
-    background-color: white;
-    border-radius: 10px;
-    
-    display: flex;
-    justify-content: center;
-    align-items: center;
-`
-
-export const ImgLogo = styled(Image)`
-  width: 100;
-  height: 100;
-`;
